+++ src/components/Core.tsx
