+++ src/components/index.tsx
@@ -1,1 +1,1 @@
-export * from './Modal'
+export * from './Core'
